Extract Navbar
